refactor: 나의 보따리 관련 파일 패키지 변경
