+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/bottari/BottariItem.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.home.team
+package com.bottari.presentation.compose.home.bottari
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
