+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/bottari/BottariMenuDropdown.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.home.team
+package com.bottari.presentation.compose.home.bottari
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
