+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/bottari/MyBottariContent.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.home.team
+package com.bottari.presentation.compose.home.bottari
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
@@ -19,7 +19,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import com.bottari.presentation.R
 import com.bottari.presentation.compose.common.component.BottariTabBar
 import com.bottari.presentation.compose.common.theme.BottariTheme
@@ -113,14 +112,7 @@ fun MyBottariContent(
             enter = fadeIn(),
             exit = fadeOut(),
         ) {
-            AddBottariButton(
-                buttonSize = 80.dp,
-                isExpanded = isFabExpanded,
-                onExpandClick = { isFabExpanded = !isFabExpanded },
-                onCodeClick = onOpenCodeDialog,
-                onTeamClick = onOpenTeamDialog,
-                onPersonalClick = onOpenPersonalDialog,
-            )
+            BottariCreateFAB()
         }
     }
 }
