+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/bottari/MyBottariDialogType.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.home.team
+package com.bottari.presentation.compose.home.bottari
 
 enum class MyBottariDialogType {
     PERSONAL,
