+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/home/bottari/MyBottariUiState.kt
@@ -1,4 +1,4 @@
-package com.bottari.presentation.compose.home.team
+package com.bottari.presentation.compose.home.bottari
 
 import com.bottari.presentation.model.bottari.MyBottariUiModel
 import com.bottari.presentation.model.bottari.personal.BottariUiModel
